refactor(admin): extract RoleBadge/UserRow/UserExpandedPanel/AssignRoleForm from UsersPanel (Related to #1852)
- UsersPanel.tsx: 717 → 203 lines, adopts useDebouncedSearch hook
- RoleBadge.tsx: colour-mapped role pill primitive
- UserRow.tsx: clickable table row + expand toggle
- UserExpandedPanel.tsx: role detail list + inline revoke confirm + AssignRoleForm
- AssignRoleForm.tsx: role assignment with scope-aware org/school picker
- UsersPanel.test.tsx: 10 TDD tests covering debounce/expand/revoke/scope-mapping
